move CompiledBinariesReport to the correct directory
