+++ varats/varats/data/reports/compiled_binary_report.py
